requirements.txt
langchain
langchain-openai
langgraph
faiss-cpu
fastapi
uvicorn
openai
pandas
numpy
python-dotenv
tiktoken
pydantic